+++ packages/magic_ext_oauth/lib/types/oauth_method.dart
@@ -1,0 +1,9 @@
+enum OAuthMethod {
+  magic_oauth_parse_redirect_result
+}
+
+extension ParseOAuthMethodToString on OAuthMethod {
+  String toShortString() {
+    return toString().split('.').last;
+  }
+}