+++ packages/magic_sdk/lib/modules/auth/auth_method.dart
@@ -1,3 +1,4 @@
+/// Auth Method
 enum AuthMethod {
   magic_auth_login_with_magic_link
 }