+++ packages/magic_sdk/lib/modules/base_module.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
-
-import 'package:magic_sdk/provider/rpc_provider.dart';
-import 'package:magic_sdk/provider/types/rpc_request.dart';
 import 'package:webview_flutter/webview_flutter.dart';
+
+import '../../provider/rpc_provider.dart';
+import '../../provider/types/rpc_request.dart';
 
 class BaseModule {
   late RpcProvider provider;
@@ -12,9 +12,9 @@ class BaseModule {
   /// Returns <JavascriptMessage> here to decode in the last step.
   /// As Dart doesn't support method return type inference. So we need to surface raw message and
   /// let the type in interface call to deserialize
-  Future<JavascriptMessage> sendToProvider({required String method, dynamic params}) async {
+  Future<JavascriptMessage> sendToProvider({required String method, List<dynamic>? params}) async {
 
-    MagicRPCRequest request = MagicRPCRequest(method: method, params: [params]);
+    MagicRPCRequest request = MagicRPCRequest(method: method, params: params ?? []);
     return provider.send(request: request, completer: Completer<JavascriptMessage>());
   }
 }