+++ packages/magic_sdk/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.3.0
+
+* Add support for Social Login
+
+## 0.2.0
+
+* Magic SDK supports web3dart
+    * sendTransaction
+    * getAccount
+    * Contract
+
 ## 0.1.0
 
-* Magic SDK Core release, 
+* Magic SDK Core release